roomid on diconnect

+++ serverless-chat/functions/disconnect/app.py
@@ -14,13 +14,31 @@ def lambda_handler(event, context):
 
     connection_id = event["requestContext"]["connectionId"]
 
-    connections_table.delete_item(
-        Key={
-            "connectionId": connection_id
+    # Find the connection
+    response = connections_table.scan(
+        FilterExpression="connectionId = :connection_id",
+        ExpressionAttributeValues={
+            ":connection_id": connection_id
         }
     )
 
-    print(f"WebSocket disconnected: {connection_id}")
+    items = response.get("Items", [])
+
+    if items:
+
+        connection = items[0]
+
+        connections_table.delete_item(
+            Key={
+                "roomId": connection["roomId"],
+                "connectionId": connection_id
+            }
+        )
+
+        print(
+            f"Disconnected: {connection_id} "
+            f"from room: {connection['roomId']}"
+        )
 
     return {
         "statusCode": 200
